Crate weather

A Ratatui example that demonstrates how to render weather data using BarChart widget.

Generates random temperature data for each hour of the day and renders it as a vertical bar.

This example runs with the Ratatui library code in the branch that you are currently reading. See the latest branch for the code which works with the most recent Ratatui release.

Functions

main 🔒

render 🔒

temperature_style 🔒

create a yellow to red value based on the value (50-90)

vertical_bar 🔒

vertical_barchart 🔒

Create a vertical bar chart from the temperatures data.